requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
requests==2.31.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-telegram-bot==20.7
slack-sdk==3.27.1
gunicorn==21.2.0
asyncio==3.4.3